Moving beer.conf validation into local manager module
